+++ R/MTD.R
@@ -89,14 +89,16 @@
 #' ## measurement of one sample (i.e., represents one *ms_run*). Columns in
 #' ## that `data.frame` provide the phenotypic and experimental variables of
 #' ## each sample. The example below represents a simple experiment in which
-#' ## 3 samples (e.g. cell lines) were measured, each at two different time
-#' ## points (0 and 6 hours). In addition, one sample has the genotype *WT* and
-#' ## two *KO*. Column `"operator"` contains the initials of the researcher
-#' ## extracting the samples
+#' ## 3 samples (e.g. cell lines) were measured. Columns *time_point and
+#' ## *cell_count* provide the time point (in hours) when samples were
+#' ## collected and the number of cells per sample. In addition, one
+#' ## sample has the genotype *WT* and two *KO*. Column `"operator"` contains
+#' ## the initials of the researcher extracting the samples
 #' exp <- data.frame(
 #'     sample_name = c("S1_T1", "S1_T2", "S2_T1", "S2_T2", "S3_T1", "S3_T2"),
 #'     sample_id = c("S1", "S1", "S2", "S2", "S3", "S3"),
-#'     timepoint = c("0h", "6h", "0h", "6h", "0h", "6h"),
+#'     time_point = c(0, 6, 0, 6, 0, 6),
+#'     cell_count = c(13000, 8700, 10100, 6000, 14000, 4500),
 #'     genotype = c("WT", "WT", "KO", "KO", "KO", "KO"),
 #'     operator = c("BB", "BB", "BB", "BB", "FB", "FB"),
 #'     file_name = c("s1-t1.mzML", "s1-t2.mzML", "s2-t1.mzML", "s2-t2.mzML",
@@ -200,13 +202,18 @@
 #' ## Study variable information
 #'
 #' ## Study variables can be defined directly from the experiment `data.frame`.
-#' ## In our example we use the columns (information on) `"timepoint"`,
-#' ## `"genotype"` and `"operator"`. Importantly, the order of the provided
-#' ## `data.frame` has to match the order of the assays (and MS runs).
+#' ## In our example we use the columns (information on) `"time_point"`,
+#' ## `"genotype"`, `"cell_count"` and `"operator"`. Importantly, the row-order
+#' ## of the provided `data.frame` has to match the order of the assays (and
+#' ## MS runs).
 #' svar <- mtd_study_variables(
-#'     exp,
-#'     study_variable_columns = c("timepoint", "genotype", "operator"))
+#'     exp, groups = c("time_point", "genotype", "cell_count", "operator"))
 #' svar
+#'
+#' ## The individual columns got encoded to a `"study_variable_group"` with a
+#' ## defined (statistical) type and a datatype. Both are inferred
+#' ## automatically from the columns of the provided `data.frame`, but could
+#' ## also be manually provided through the respective function arguments.
 #'
 #' mtd <- rbind(mtd, svar)
 #'
@@ -435,12 +442,16 @@ mtd_fields <- function(..., field_prefix = "") {
 mtd_skeleton <- function(id = character(),
                          software = character(),
                          quantification_method = "[MS, MS:1001834, LC-MS label-free quantitation analysis, ]",
-                         cv_label = c("MS", "PRIDE"),
+                         cv_label = c("MS", "PRIDE", "STATO"),
                          cv_full_name = c("PSI-MS controlled vocabulary",
-                                          "PRIDE PRoteomics IDEntifications (PRIDE) database controlled vocabulary"),
-                         cv_version = c("4.1.138", "16:10:2023 11:38"),
+                                          "PRIDE PRoteomics IDEntifications (PRIDE) database controlled vocabulary",
+                                          "General purpose STATistics Ontology"),
+                         cv_version = c("4.1.138",
+                                        "16:10:2023 11:38",
+                                        "2026-04-20"),
                          cv_uri = c("https://raw.githubusercontent.com/HUPO-PSI/psi-ms-CV/master/psi-ms.obo",
-                                    "https://www.ebi.ac.uk/ols/ontologies/pride"),
+                                    "https://www.ebi.ac.uk/ols/ontologies/pride",
+                                    "https://www.ebi.ac.uk/ols4/ontologies/stato"),
                          database = c("[,, \"no database\", null ]"),
                          database_prefix = c("null"),
                          database_version = c("Unknown"),
@@ -448,7 +459,7 @@ mtd_skeleton <- function(id = character(),
                          small_molecule_quantification_unit = "[PRIDE, PRIDE:0000330, Arbitrary quantification unit, ]",
                          small_molecule_feature_quantification_unit = "[PRIDE, PRIDE:0000330, Arbitrary quantification unit, ]",
                          small_molecule_identification_reliability = "[MS, MS:1002896, compound identification confidence level, ]",
-                         mztab_version = "2.0.0-M") {
+                         mztab_version = "2.1.0-M") {
     if (!length(id)) stop("Parameter 'id' is required", call. = FALSE)
     if (!length(software)) stop("Parameter 'software' is required", call.=FALSE)
     sk <- rbind(
@@ -883,60 +894,97 @@ mtd_assay <- function(..., assay = character(), external_uri = character(),
 #'
 #' @description
 #'
-#' A *study variable* in the mzTab-M definition is the **value** of an
-#' experimental variable (condition or factor), e.g., `"control"`, or
-#' `"2 hours"` (i.e., it is not the name of the factor). Each study variable
-#' **must** be reported in the following abundance tables. Each assay of a data
-#' set must be referred to from at least one study variable* Even if a data set
-#' has no experimental variables, a study variable **must** be reported,
-#' linking to all assays, with the name `"undefined"`.
+#' mzTab-M (version >= 2.1) encodes the experimental design of a data set/study
+#' using *study variable groups* and *study variables*. The study variable group
+#' represents the phenotypic (or experimental) condition and the study variable
+#' the actual *value* of a sample (or assay) for that study group.
 #'
-#' Information for the mzTab-M *study variable* metadata section can be created
-#' using the `mtd_study_variables()` function. This function extracts the
-#' relevant information from an input `data.frame` in which rows are expected
-#' to represent *assays* (each row being one assay) and columns the experimental
-#' factors that should be added as study variables.
+#' In R, the most common representation of an experimental design is a
+#' `data.frame` where rows are individual samples (assays) and columns the
+#' experimental or phenotypic conditions (variables). The
+#' `mtb_study_variables()` takes such a `data.frame` as input and encodes it
+#' into the mzTab-M format. Additional parameters such as `group_description`,
+#' and `group_type` allow to provide additional information for each study
+#' variable group (phenotype) while parameters `average_function`,
+#' `variation_function` and `description` can be used to provide properties for
+#' the individual study variables. For most experiments the default values of
+#' these parameters should suffice.
+#'
+#' @details
+#'
+#' Each study variable **must** be reported in the abundance tables.
+#' Each assay of a data set must be referred to from at least one study
+#' variable. Even if a data set has no experimental variables, a study variable
+#' group and study variable with the name `"undefined"` **must** be reported.
+#' Using `mtb_study_variables()` without specifying parameter `group` will
+#' create such a setup.
 #'
 #' The `mtd_define_study_variables()` function can be used to get the set
 #' (and order) of study variables that would be generated from an input
-#' `data.frame` depending on the parameter `study_variable_columns`.
+#' `data.frame` depending on the parameter `groups`.
+#'
+#' @note
+#'
+#' Datatypes `"xsd:date"`, `"xsd:time"`, `"xsd:dateTime"` and `"xsd:anyURI"` are
+#' currently mapped to `character` in R (and *vice versa*).
 #'
 #' @param x `data.frame` with rows corresponding to individual *assays* and
 #'     columns containing the experimental conditions/study variables. The
 #'     number of rows is thus expected to be the same as the number of assays
 #'     defined in the *assay* metadata section (using e.g., [mtd_assay()]) and
-#'     the order of rows is expected to match the order of assays.
+#'     the order of rows is expected to match the order of these.
 #'
-#' @param study_variable_columns `character` with the names of the columns in
-#'     `x`to be included as study variables. If not defined (the default) a
-#'     single study variable will be defined assigning all assays to it.
+#' @param groups `character` with the names of the columns in `x` that should
+#'     be considered as *study variable groups*. If not defined (the default)
+#'     a single study variable group `"undefined"` and single study variable
+#'     `"undefined"` will be used.
+#'
+#' @param group_description `character` with an optional description of each
+#'     study variable group. If provided its length has to match the length
+#'     of parameter `groups`.
+#'
+#' @param group_type `character` defining the type for each study variable
+#'     group. If provided its length has to match the length of parameter
+#'     `groups`. Supported values are
+#'     `"[STATO, STATO:0000252, categorical variable]"`,
+#'     `"[STATO, STATO:0000228, ordinal variable]"` and
+#'     `"[STATO, STATO:0000251, continuous variable]"` for categorical, ordinal
+#'     or numerical values, respectively. If not provided (the default) the
+#'     study variable group type will be inferred from the data type of the
+#'     respective columns in `x`.
+#'
+#' @param group_datatype optional `character` defining the data type of the
+#'     values (i.e., study variables) for the study variable group. If provided,
+#'     its length has to match the length of paramter `groups`.
+#'     Supported values are `"xsd:string"`, `"xsd:integer"`, `"xsd:decimal"`,
+#'     `"xsd:boolean"`, `"xsd:date"`, `"xsd:time"`, `"xsd:dateTime"`, and
+#'     `"xsd:anyURI"`. Date,l time and dateTime values **must** be encoded in
+#'     ISO 8601 format. If not provided the type is guessed by the data type of
+#'     the respective column in `x`.
 #'
 #' @param average_function optional `character` defining the function used to
 #'     calculate the study variable quantification value (reported in the
 #'     following table(s)). Can be of length 1 or equal to the number of study
 #'     variables (to allow defining a different function per variable). Use
 #'     `mtd_define_study_variables()` to get the complete set of study
-#'     variables for parameters `x` and `study_variable_columns`. Defaults
-#'     to the arithmetic mean
-#'     (`average_function = "[MS, MS:1002962, mean, ]"`).
+#'     variables for parameters `x` and `groups`. Defaults to the arithmetic
+#'     mean (`average_function = "[MS, MS:1002962, mean, ]"`).
 #'
 #' @param variation_function optional `character` defining the function used to
 #'     calculate the study variable quantification variation value (reported in
 #'     the following table(s)). Can be of length 1 or equal to the number of
 #'     study variables (to allow defining a different function per variable).
 #'     Use `mtd_define_study_variables()` to get the complete set of study
-#'     variables for parameters `x` and `study_variable_columns`. Defaults
+#'     variables for parameters `x` and `groups`. Defaults
 #'     to the coefficient of variation
 #'     (`variation_function = "[MS, MS:1002963, variation coefficient, ]"`).
 #'
 #' @param description `character` with a textual description of the study
 #'     variable. If provided, its length needs to be equal to the number of
 #'     study variables. Use `mtd_define_study_variables()` to get the complete
-#'     set of study variables for parameters `x` and `study_variable_columns`.
-#'     If not provided (the default) the description is used combines the
-#'     column name in `x` and the value of the variable.
-#'
-#' @param factors Currently not supported.
+#'     set of study variables for parameters `x` and `groups`.
+#'     If not provided (the default) the values for the study variable group
+#'     and study variable are reported.
 #'
 #' @return two-column `character` `matrix` with the content for the study
 #'     variables metadata section.
@@ -955,64 +1003,74 @@ mtd_assay <- function(..., assay = character(), external_uri = character(),
 #' x <- data.frame(
 #'     name = c("I1_0", "I2_0", "I1_6", "I2_6", "I3_0"),
 #'     individual = c("I1", "I2", "I1", "I2", "I3"),
-#'     timepoint = c("0h", "6h", "0h", "6h", "0h"),
+#'     BMI = c(29.3, 31.4, 29.3, 31.4, 26.5),
+#'     timepoint = c(0, 6, 0, 6, 0),
 #'     T2D = c(TRUE, FALSE, TRUE, FALSE, FALSE)
 #' )
 #'
-#' ## Study variables for this data set would be `"timepoint"` and `"T2D"`:
-#' mtd_study_variables(x, study_variable_columns = c("timepoint", "T2D"))
+#' ## Study variable groups for this data set could be `"individual"`, `"BMI"`,
+#' ## `"timepoint"` and `"T2D"`
+#' mtd_study_variables(x, groups = c("individual", "BMI", "timepoint", "T2D"))
 #'
-#' ## Specifying a different average and variation function
+#' ## Specifying a different average and variation function and selecting
+#' ## just two sample columns
 #' mtd_study_variables(x,
-#'     study_variable_columns = c("timepoint", "T2D"),
+#'     groups = c("timepoint", "T2D"),
 #'     average_function = "[MS, MS:1002883, median, ]",
 #'     variation_function = "[MS, MS:1002885, standard error, ]")
 #'
-#' ## Creating a study variable section without defined study variables
+#' ## Creating a study variable section without defined study variable groups
 #' mtd_study_variables(x)
-mtd_study_variables <- function(x, study_variable_columns = character(),
+#'
+#' ## Use `mtd_define_study_variables()` to get the definition of study
+#' ## variables for a given `x` and `groups`
+#' mtd_define_study_variables(x, c("T2D", "BMI", "individual"))
+mtd_study_variables <- function(x, groups = character(),
+                                group_description = character(),
+                                group_type = character(),
+                                group_datatype = character(),
                                 average_function = "[MS, MS:1002962, mean, ]",
                                 variation_function = "[MS, MS:1002963, variation coefficient, ]",
-                                description = character(),
-                                factors = character()) {
-    if (length(factors))
-        stop("'factors' is currently not supported", call. = FALSE)
-    if (length(study_variable_columns)) {
-        if (!all(study_variable_columns %in% colnames(x)))
-            stop("Not all column names defined with 'study_variable_columns' ",
-                 "present in 'x'", call. = FALSE)
-        svar_m <- .mztab_study_variables(x, study_variable_columns)
-        svars <- unique(as.vector(svar_m))
-    } else {
-        ## Define a single study variable and assign all assays to it.
-        svars <- "undefined"
-        svar_m <- matrix(ncol = 1, nrow = nrow(x), svars)
-        description <- "Undefined"
+                                description = character()) {
+    if (!length(groups)) {
+        x <- data.frame(undefined = rep("undefined", nrow(x)))
+        groups <- "undefined"
     }
-    l <- length(svars)
-    if (!length(description)) {
-        description <- vapply(svars, function(z) {
-            z <- strsplit(z, split = ":")[[1L]]
-            paste0("Column: ", z[1L], ", value: ", z[2L])
-        }, NA_character_)
-    }
+    if (!all(groups %in% colnames(x)))
+        stop("Not all column names defined with 'groups' are ",
+             "present in 'x'", call. = FALSE)
+    x <- as.data.frame(x[, groups, drop = FALSE])
+    svg <- c(groups, .mtd_svar_group_description(x, group_description),
+             .mtd_svar_group_type(x, group_type),
+             .mtd_svar_group_datatype(x, group_datatype))
+    ## build study variable group content
+    res <- cbind(paste0("study_variable_group[",
+                        rep(seq_along(groups), each = 4L),
+                        c("]", "]-description", "]-type", "]-datatype")),
+                 svg[order(rep(seq_along(groups), 4L))])
+    svar_df <- .mztab_study_variables(x, groups)
+    svars <- unique(svar_df)
+    l <- nrow(svars)
     if (length(average_function) == 1L)
         average_function <- rep(average_function, l)
     if (length(variation_function) == 1L)
         variation_function <- rep(variation_function, l)
-    if (length(average_function)  != l)
+    if (length(average_function) != l)
         stop("Length of parameter 'average_function' has to be equal to ",
              "the number of study variables", call. = FALSE)
-    if (length(variation_function)  != l)
+    if (length(variation_function) != l)
         stop("Length of parameter 'variation_function' has to be equal to ",
              "the number of study variables", call. = FALSE)
-    if (length(description)  != l)
+    if (!length(description))
+        description <- paste0("Variable ", svars$study_variable_group,
+                              ", value ", svars$study_variable)
+    if (length(description) != l)
         stop("Length of parameter 'description' has to be equal to ",
              "the number of study variables", call. = FALSE)
-    ## Build matrix
-    res <- matrix(ncol = 2, nrow = 0, NA_character_)
-    for (i in seq_along(svars)) {
-        idx <- which(svar_m == svars[i], arr.ind = TRUE)
+    ## Add study variables
+    for (i in seq_len(nrow(svars))) {
+        current_svar <- svars$study_variable[i]
+        current_grp <- svars$study_variable_group[i]
         res <- rbind(
             res,
             matrix(ncol = 2,
@@ -1021,11 +1079,17 @@ mtd_study_variables <- function(x, study_variable_columns = character(),
                      paste0("study_variable[", i, "]-average_function"),
                      paste0("study_variable[", i, "]-variation_function"),
                      paste0("study_variable[", i, "]-description"),
-                     paste0(svars[i]),
-                     paste0("assay[", idx[, "row"], "]", collapse = "|"),
+                     paste0("study_variable[", i, "]-group_refs"),
+                     current_svar,
+                     paste0("assay[", which(x[, current_grp] == current_svar),
+                            "]", collapse = "|"),
                      average_function[i],
                      variation_function[i],
-                     description[i])))
+                     description[i],
+                     paste0("study_variable_group[",
+                            match(current_grp, groups), "]"))
+                   )
+        )
     }
     res
 }
@@ -1033,8 +1097,14 @@ mtd_study_variables <- function(x, study_variable_columns = character(),
 #' @rdname mtd_study_variables
 #'
 #' @export
-mtd_define_study_variables <- function(x, study_variable_columns = character()){
-    unique(as.vector(.mztab_study_variables(x, study_variable_columns)))
+mtd_define_study_variables <- function(x = data.frame(), groups = character()) {
+    if (!nrow(x))
+        return(data.frame(study_variable = character(),
+                          study_variable_group = character()))
+    if (!length(groups))
+        data.frame(study_variable = "undefined",
+                   study_variable_group = "undefined")
+    else unique(.mztab_study_variables(x, groups))
 }
 
 #' @title Sort rows in a MTD matrix to match the expected order
@@ -1217,17 +1287,15 @@ mtd_sort <- function(x) {
 
 #' @description
 #'
-#' Creates a `matrix` of study variables.
+#' Creates a `data.frame` of study variables and their study variable groups.
 #'
 #' @param x `data.frame` with sample annotations. Each row being one assay.
 #'
-#' @param study_variable_columns `character` with the names of the column(s)
+#' @param groups `character` with the names of the column(s)
 #'     from which study variables should be defined.
 #'
-#' @return `character` `matrix`, same number of rows as `x` with the study
-#'     variables in columns. Each cell represents the study variable value
-#'     allowing to assign a specific sample (row) to the respective study
-#'     variable.
+#' @return `data.frame` with all study variable (values) in one column and the
+#'     associated study variable group (column) in the second.
 #'
 #' @noRd
 #'
@@ -1242,10 +1310,11 @@ mtd_sort <- function(x) {
 #'
 #' .mztab_study_variables(x, "sex")
 .mztab_study_variables <- function(x = data.frame(),
-                                   study_variable_columns = colnames(x),
-                                   sep = ":") {
-    do.call(cbind, lapply(study_variable_columns,
-                          function(z) paste0(z, sep, x[, z])))
+                                   groups = colnames(x)) {
+    data.frame(
+        study_variable = as.vector(unlist(lapply(x[, groups, drop = FALSE],
+                                                 as.character))),
+        study_variable_group = rep(groups, each = nrow(x)))
 }
 
 #' Defines the order of the elements in MTD (pattern provided). This should
@@ -1268,6 +1337,7 @@ mtd_sort <- function(x) {
     "^sample",
     "^ms_run",
     "^assay",
+    "^study_variable_group",
     "^study_variable",
     "^custom",
     "^cv",
@@ -1360,3 +1430,118 @@ mtd_sort <- function(x) {
         else NA_character_
     })
 }
+
+#' Helper function to check and define the study_variable_group_description.
+#'
+#' @param x `data.frame` with columns used as study_variable_group
+#'
+#' @return `character` of length equal to `ncol(x)` with the study variable
+#'     group description.
+#'
+#' @noRd
+.mtd_svar_group_description <- function(x, group_description = character()) {
+    grp <- colnames(x)
+    if (length(group_description)) {
+        if (length(grp) != length(group_description))
+            stop("Length of 'group_description' has to match length of ",
+                 "'groups'.", call. = FALSE)
+        group_description
+    } else paste0("Sample matrix column ", grp)
+}
+
+#' Helper function to check or guess the (STATO) study variable group type
+#' from the R data types of columns in the input `data.frame` `x`.
+#'
+#' @param x `data.frame` with columns being the study variable types
+#'
+#' @param group_type `character`, if provided, of length equal to `ncol(x)` with
+#'     the data types to check for validity of the provided names. If not
+#'     provided, the data type is guessed on the data type of the column.
+#'
+#' @return `character` with the param of the data type
+#'
+#' @noRd
+.mtd_svar_group_type <- function(x, group_type = character()) {
+    if (length(group_type)) {
+        if (length(group_type) != ncol(x))
+            stop("If provided, the length of 'group_type' has to match the ",
+                 "number of study variable groups.", call. = FALSE)
+        ## Check that it contains supported types.
+        st <- parse_cv_parameter(group_type)
+        idx <- match(st, parse_cv_parameter(.STUDY_VARIABLE_GROUP_TYPE$stato))
+        if (anyNA(idx)) {
+            stop("Group types ", paste0("\"", st[is.na(idx)], "\"",
+                                        collapse = ", "),
+                 " are not supported. See ?mtd_study_variables for supported",
+                 " study variable group types.", call. = FALSE)
+        }
+        .STUDY_VARIABLE_GROUP_TYPE$stato[idx]
+    } else {
+        dtype <- vapply(x, function(z) class(z)[1L], NA_character_)
+        ## get FIRST match for each
+        idx <- match(dtype, .STUDY_VARIABLE_GROUP_TYPE$r)
+        if (anyNA(idx))
+            stop("Group type(s) ", paste0(dtype[is.na(idx)], collapse = ", "),
+                 " are not supported", call. = TRUE)
+        .STUDY_VARIABLE_GROUP_TYPE$stato[idx]
+    }
+}
+
+#' Mapping of study variable type to R data types.
+#'
+#' STATO:0000252: categorical
+#' STATO:0000251: continuous
+#' STATO:0000228: ordinal
+#'
+#' @noRd
+.STUDY_VARIABLE_GROUP_TYPE <- data.frame(
+    r = c("character", "factor", "logical", "numeric", "integer", "integer"),
+    stato = c("[STATO, STATO:0000252, categorical variable]",
+              "[STATO, STATO:0000252, categorical variable]",
+              "[STATO, STATO:0000252, categorical variable]",
+              "[STATO, STATO:0000251, continuous variable]",
+              "[STATO, STATO:0000251, continuous variable]",
+              "[STATO, STATO:0000228, ordinal variable]")
+)
+
+#' Helper function to check or define the datatype for study variable groups.
+#'
+#' @param x `data.frame` with columns being the study variable groups
+#'
+#' @param group_datatype `character` with the datatype for the groups (columns).
+#'     If provided, it will be checked for valid entries, if not provided it
+#'     will be guessed from the data type of the columns.
+#'
+#' @return `character` of length equal to `ncol(x)` with the datatype for each
+#'     column.
+#'
+#' @noRd
+.mtd_svar_group_datatype <- function(x, group_datatype = character()) {
+    if (length(group_datatype)) {
+        if (length(group_datatype) != ncol(x))
+            stop("If provided, the length of 'group_datatype' has to match ",
+                 "the number of study variable groups.", call. = FALSE)
+        idx <- match(group_datatype, .STUDY_VARIABLE_GROUP_DATATYPE$xsd)
+        if (anyNA(idx)) {
+            stop("Group datatypes ",
+                 paste0("\"", group_datatype[is.na(idx)], "\"",collapse = ", "),
+                 " are not supported. See ?mtd_study_variables for supported",
+                 " study variable group datatypes.", call. = FALSE)
+        }
+        .STUDY_VARIABLE_GROUP_DATATYPE$xsd[idx]
+    } else {
+        dtype <- vapply(x, function(z) class(z)[1L], NA_character_)
+        idx <- match(dtype, .STUDY_VARIABLE_GROUP_DATATYPE$r)
+        if (anyNA(idx))
+            stop("Group datatype(s) ", paste0(dtype[is.na(idx)],collapse =", "),
+                 " are not supported", call. = TRUE)
+        .STUDY_VARIABLE_GROUP_DATATYPE$xsd[idx]
+    }
+}
+
+.STUDY_VARIABLE_GROUP_DATATYPE <- data.frame(
+    r = c("character", "integer", "numeric", "logical", "factor",
+          "character", "character", "character", "character"),
+    xsd = c("xsd:string", "xsd:integer", "xsd:decimal", "xsd:boolean",
+            "xsd:string", "xsd:dateTime", "xsd:date", "xsd:time", "xsd:anyURI")
+)